Moved Stage Instance related APIs to StageInstanceControllerV2
 - updated the stage instance API to following end point:
    `api/stage/:pipeline_name/:pipeline_counter/:stage_name/:stage_counter`
