+++ spark/spark-base/src/main/java/com/thoughtworks/go/spark/Routes.java
@@ -327,6 +327,7 @@ public class Routes {
         public static final String TRIGGER_SELECTED_JOBS_PATH = "/:pipeline_name/:pipeline_counter/:stage_name/:stage_counter/run-selected-jobs";
         public static final String CANCEL_STAGE_PATH = "/:pipeline_name/:pipeline_counter/:stage_name/:stage_counter/cancel";
         public static final String INSTANCE_BY_COUNTER = "/:pipeline_name/:stage_name/instance/:pipeline_counter/:stage_counter";
+        public static final String INSTANCE_V2 = "/:pipeline_name/:pipeline_counter/:stage_name/:stage_counter";
         public static final String STAGE_HISTORY = "/:pipeline_name/:stage_name/history";
         public static final String STAGE_HISTORY_OFFSET = "/:pipeline_name/:stage_name/history/:offset";
